Move business logic to services
